Move manual update cases
Note: The test
  test_azure_instance_password_requirements_azure
is a dup with
  test_azure_instance_password_requirements

Note: The test
  test_vm_placement_with_duplicated_folder_name_vmware
is already in
  cfme/tests/cloud_infra_common/test_provisioning_manual.py

Note: The test
    test_appliance_log_error
is already in
    cfme/tests/test_appliance.py
