+++ cfme/tests/configure/test_proxy.py
@@ -181,3 +181,345 @@ def test_proxy_invalid(appliance, prepare_proxy_invalid, provider):
             view.entities.summary('Status').get_text_of('Last Refresh'))
 
     wait_for(last_refresh_failed, fail_condition=False, num_sec=240, delay=5)
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_remove_default():
+    """
+    With 5.7 there is a new feature that allows users to specific a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the default proxy settings.  For this test you want to create
+    an default proxy, verified it worked, and then remove the proxy and
+    verify it didn"t use a proxy
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        initialEstimate: 1/8h
+        setup: I think the best way to do this one is start with a bad proxy value,
+               get the connection error, and then remove the proxy values and make
+               sure it starts connecting again.  I"ll have to see if there is a log
+               value we can look at.  Otherwise, you need to shutdown the proxy
+               server to be absolutely sure.
+        startsin: 5.7
+        upstream: yes
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_remove_azure():
+    """
+    With 5.7 there is a new feature that allows users to specific a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the azure proxy settings.  For this test you want to create an
+    azure proxy, verified it worked, and then remove the proxy and verify
+    it used the default which may or may not be blank.
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        initialEstimate: 1/8h
+        setup: I think the best way to do this one is start with a bad proxy value,
+               get the connection error, and then remove the proxy values and make
+               sure it starts connecting again.  I"ll have to see if there is a log
+               value we can look at.  Otherwise, you need to shutdown the proxy
+               server to be absolutely sure.
+        startsin: 5.7
+        upstream: yes
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_override_gce():
+    """
+    With 5.7 there is a new feature that allows users to specific a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the ec2 proxy settings.  For this test you want to create a
+    bogus setting for the default entry and a correct entry for ec2 so
+    that you can make sure it is switching to ec2 correctly.
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        initialEstimate: 1/8h
+        setup: Make sure you create a bad proxy for default and a correct proxy for
+               ec2 so that you are certain we grab the right entry.
+        startsin: 5.7
+        upstream: yes
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_override_ec2():
+    """
+    With 5.7 there is a new feature that allows users to specific a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the ec2 proxy settings.  For this test you want to create a
+    bogus setting for the default entry and a correct entry for ec2 so
+    that you can make sure it is switching to ec2 correctly.
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        initialEstimate: 1/8h
+        setup: Make sure you create a bad proxy for default and a correct proxy for
+               ec2 so that you are certain we grab the right entry.
+        startsin: 5.7
+        upstream: yes
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_override_azure():
+    """
+    With 5.7 there is a new feature that allows users to specific a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the azure proxy settings.  For this test you want to create a
+    bogus setting for the default entry and a correct entry for azure so
+    that you can make sure it is switch to azure correctly.
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        initialEstimate: 1/4h
+        setup: Setup the azure proxy correct and the default proxy incorrectly and
+               make sure azure uses the correct entry.
+        startsin: 5.7
+        upstream: yes
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_invalid_azure():
+    """
+    With 5.7 there is a new feature that allows users to specific a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the azure proxy settings.  You just need to fill in the
+    appropriate information, only screw up a few of the values to make
+    sure it reports a connection error.
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        caseposneg: negative
+        initialEstimate: 1/2h
+        upstream: yes
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_invalid_gce():
+    """
+    With 5.7 there is a new feature that allows users to specific a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the azure proxy settings.  You just need to fill in the
+    appropriate information, only screw up a few of the values to make
+    sure it reports a connection error.
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        caseposneg: negative
+        initialEstimate: 1/8h
+        setup: The only thing different you need to do for this is enter some wrong
+               information and Refresh Relationships.  Wait two minutes and refresh
+               the page.  You"ll definitely get an error if any of the values are
+               wrong.
+        startsin: 5.7
+        upstream: yes
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_invalid_default():
+    """
+    With 5.7 there is a new feature that allows users to specify a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the default proxy settings.  You just need to fill in the
+    appropriate information, only screw up a few of the values to make
+    sure it reports a connection error.
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        caseposneg: negative
+        initialEstimate: 1/2h
+        setup: The mojo page has all the information you will need.
+        startsin: 5.7
+        upstream: yes
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_invalid_ec2():
+    """
+    With 5.7 there is a new feature that allows users to specify a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the ec2 proxy settings.  You just need to fill in the
+    appropriate information, only screw up a few of the values to make
+    sure it reports a connection error.
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        caseposneg: negative
+        initialEstimate: 1/4h
+        setup: Follow the instructions in the mojo document.
+        startsin: 5.7
+        teardown: You should probably reset or delete the changes.
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_valid_gce():
+    """
+    With 5.7.1 there is a new feature that allows users to specific a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the azure proxy settings.  You just need to fill in the
+    appropriate information
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        initialEstimate: 1/4h
+        setup: Configure advanced settings for GCE proxy.
+               Add gce provider
+               Probably need to check the packets to make sure they are vectoring
+               through the proxy server, or just check the proxy server log.
+        startsin: 5.7
+        upstream: yes
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_valid_default():
+    """
+    With 5.7 there is a new feature that allows users to specific a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the azure proxy settings.  You just need to fill in the
+    appropriate information
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        initialEstimate: 1/2h
+        setup: Follow the instructions in the mojo doc above.
+        startsin: 5.7
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_valid_azure():
+    """
+    With 5.7 there is a new feature that allows users to specific a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the azure proxy settings.  You just need to fill in the
+    appropriate information
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        initialEstimate: 1/2h
+        setup: Configure advanced settings for Azure proxy.
+               Add azure provider
+               Probably need to check the packets to make sure they are vectoring
+               through the proxy server, or just check the proxy server log.
+        startsin: 5.7
+        upstream: yes
+    """
+    pass
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_proxy_valid_ec2():
+    """
+    With 5.7 there is a new feature that allows users to specify a
+    specific set of proxy settings for each cloud provider.  The way you
+    enable this is to go to Configuration/Advanced Settings and scroll
+    down to the ec2 proxy settings.  You just need to fill in the
+    appropriate information
+    Here are the proxy instructions:
+    https://mojo.redhat.com/docs/DOC-1103999
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Cloud
+        caseimportance: medium
+        initialEstimate: 1/2h
+        setup: Follow the instructions in the mojo doc above as it is easier to
+               change in one place.
+        startsin: 5.7
+    """
+    pass
