+++ cfme/tests/platform/test_locale.py
@@ -1,0 +1,17 @@
+import pytest
+
+
+@pytest.mark.manual
+@pytest.mark.tier(1)
+def test_automated_locale_switching():
+    """
+    Having the automatic locale selection selected, the appliance"s locale
+    changes accordingly with user"s preferred locale in the browser.
+
+    Polarion:
+        assignee: jhenner
+        casecomponent: Appliance
+        caseimportance: medium
+        initialEstimate: 1/8h
+    """
+    pass
